requirements.txt
Faker==18.9.0
mysql-connector-python==8.3.0
